Ignore calls when stopped on the floor
When an elevator is stopped on the floor the request came from, and it's
already going the direction requested, it should ignore the call.
Otherwise it will come back and service it later. But the passenger has
already gone the way they wanted.

Doing this required tracking a new piece of state: the elevator's
current floor.